suprression de spiff et modifs

+++ AMOC_TP2_Reveil_Matin/include/AlarmMelodies.h
@@ -10,8 +10,8 @@ private:
     uint8_t m_nbSnoozes;
     uint8_t m_melodyIndex;
     Buzzer* m_buzzer;
-    int16_t m_delay1;
-    int16_t m_delay2;
+    uint64_t m_delay1;
+    uint64_t m_delay2;
     uint64_t m_lastUpdate;
     int16_t* m_melodyArray[NB_MELODIES];
     size_t m_melodyLength[NB_MELODIES];
@@ -21,8 +21,8 @@ private:
 
 public:
     AlarmMelodies(
-        uint16_t p_delay1,
-        uint16_t p_delay2,
+        uint64_t p_delay1,
+        uint64_t p_delay2,
         Buzzer* p_buzzer
         );
     void activateAlarm();
